Move Wasm encoding implementations to hermes-wasm-encoding-components crate
