+++ crates/cosmos/cosmos-wasm-relayer/src/encoding/convert.rs
@@ -4,11 +4,12 @@ use hermes_cosmos_chain_components::types::tendermint::{
     ProtoTendermintClientState, ProtoTendermintConsensusState, TendermintClientState,
     TendermintConsensusState,
 };
-use hermes_wasm_client_components::impls::encoding::components::WasmEncodingComponents;
-use hermes_wasm_client_components::types::client_state::WasmClientState;
-use hermes_wasm_client_components::types::consensus_state::{
-    DecodeViaWasmConsensusState, EncodeViaWasmConsensusState, WasmConsensusState,
+use hermes_wasm_encoding_components::components::WasmEncodingComponents;
+use hermes_wasm_encoding_components::impls::strategies::consensus_state::{
+    DecodeViaWasmConsensusState, EncodeViaWasmConsensusState,
 };
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 use ibc::core::commitment_types::merkle::MerkleProof;
 use ibc_proto::ibc::core::commitment::v1::MerkleProof as ProtoMerkleProof;
 use prost_types::Any;
