+++ crates/cosmos/cosmos-wasm-relayer/src/encoding/encode_mut.rs
@@ -1,8 +1,8 @@
 use cgp::prelude::*;
 use hermes_protobuf_encoding_components::types::strategy::ViaProtobuf;
-use hermes_wasm_client_components::impls::encoding::components::WasmEncodingComponents;
-use hermes_wasm_client_components::types::client_state::WasmClientState;
-use hermes_wasm_client_components::types::consensus_state::WasmConsensusState;
+use hermes_wasm_encoding_components::components::WasmEncodingComponents;
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 use ibc::core::client::types::Height;
 
 pub struct WasmCosmosEncodeMutComponents;
