+++ crates/cosmos/cosmos-wasm-relayer/src/encoding/type_url.rs
@@ -3,9 +3,9 @@ use hermes_cosmos_chain_components::encoding::components::CosmosClientEncodingCo
 use hermes_cosmos_chain_components::types::tendermint::{
     TendermintClientState, TendermintConsensusState,
 };
-use hermes_wasm_client_components::impls::encoding::components::WasmEncodingComponents;
-use hermes_wasm_client_components::types::client_state::WasmClientState;
-use hermes_wasm_client_components::types::consensus_state::WasmConsensusState;
+use hermes_wasm_encoding_components::components::WasmEncodingComponents;
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 
 pub struct WasmCosmosTypeUrlSchemas;
