+++ crates/cosmos/cosmos-wasm-relayer/src/impls/create_client_message.rs
@@ -16,8 +16,8 @@ use hermes_relayer_components::chain::traits::types::create_client::{
     HasCreateClientMessageOptionsType, HasCreateClientPayloadType,
 };
 use hermes_relayer_components::chain::traits::types::message::HasMessageType;
-use hermes_wasm_client_components::types::client_state::WasmClientState;
-use hermes_wasm_client_components::types::consensus_state::WasmConsensusState;
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 use ibc::core::client::types::Height;
 use prost_types::Any;
