+++ crates/wasm/wasm-client-components/Cargo.toml
@@ -19,6 +19,7 @@ hermes-relayer-components               = { workspace = true }
 hermes-encoding-components              = { workspace = true }
 hermes-protobuf-encoding-components     = { workspace = true }
 hermes-cosmos-encoding-components       = { workspace = true }
+hermes-wasm-encoding-components         = { workspace = true }
 
 ibc-relayer-types                       = { workspace = true }
 tendermint-proto                        = { workspace = true }
