--- crates/wasm/wasm-client-components/src/impls/encoding/components.rs
@@ -1,47 +1,0 @@
-use cgp::prelude::*;
-pub use hermes_cosmos_chain_components::encoding::components::{
-    DecodeBufferTypeComponent, EncodeBufferTypeComponent,
-};
-use hermes_cosmos_encoding_components::components::CosmosEncodingComponents;
-use hermes_encoding_components::impls::delegate::DelegateEncoding;
-pub use hermes_encoding_components::traits::convert::ConverterComponent;
-pub use hermes_encoding_components::traits::decode::DecoderComponent;
-pub use hermes_encoding_components::traits::decode_mut::MutDecoderComponent;
-pub use hermes_encoding_components::traits::encode::EncoderComponent;
-pub use hermes_encoding_components::traits::encode_mut::MutEncoderComponent;
-pub use hermes_encoding_components::traits::schema::SchemaGetterComponent;
-pub use hermes_encoding_components::traits::types::encoded::EncodedTypeComponent;
-pub use hermes_encoding_components::traits::types::schema::SchemaTypeComponent;
-pub use hermes_protobuf_encoding_components::traits::length::EncodedLengthGetterComponent;
-
-use crate::impls::encoding::convert::WasmConverterComponents;
-use crate::impls::encoding::encode_mut::WasmEncodeMutComponents;
-use crate::impls::encoding::encoder::WasmEncoderComponents;
-use crate::impls::encoding::type_url::WasmTypeUrlSchemas;
-
-define_components! {
-    WasmEncodingComponents {
-        [
-            EncodedTypeComponent,
-            EncodeBufferTypeComponent,
-            DecodeBufferTypeComponent,
-            SchemaTypeComponent,
-        ]:
-            CosmosEncodingComponents,
-        ConverterComponent:
-            DelegateEncoding<WasmConverterComponents>,
-        [
-            EncoderComponent,
-            DecoderComponent,
-        ]:
-            DelegateEncoding<WasmEncoderComponents>,
-        [
-            MutEncoderComponent,
-            MutDecoderComponent,
-            EncodedLengthGetterComponent,
-        ]:
-            DelegateEncoding<WasmEncodeMutComponents>,
-        SchemaGetterComponent:
-            DelegateEncoding<WasmTypeUrlSchemas>,
-    }
-}
