+++ crates/wasm/wasm-client-components/src/impls/types/client_state.rs
@@ -1,7 +1,6 @@
 use cgp::core::Async;
 use hermes_relayer_components::chain::traits::types::client_state::ProvideClientStateType;
-
-use crate::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
 
 pub struct ProvideWasmClientState;
