+++ crates/wasm/wasm-client-components/src/lib.rs
@@ -2,4 +2,3 @@
 
 pub mod contexts;
 pub mod impls;
-pub mod types;
