+++ crates/wasm/wasm-encoding-components/src/impls/mod.rs
@@ -1,0 +1,2 @@
+pub mod encode;
+pub mod strategies;
