+++ crates/wasm/wasm-encoding-components/src/impls/strategies/consensus_state.rs
@@ -1,47 +1,11 @@
-use cgp::prelude::*;
-use hermes_encoding_components::impls::encode_mut::field::EncodeField;
-use hermes_encoding_components::impls::encode_mut::from::DecodeFrom;
 use hermes_encoding_components::traits::convert::{CanConvert, Converter};
 use hermes_encoding_components::traits::decode::{CanDecode, Decoder};
-use hermes_encoding_components::traits::decode_mut::MutDecoderComponent;
 use hermes_encoding_components::traits::encode::{CanEncode, Encoder};
-use hermes_encoding_components::traits::encode_mut::MutEncoderComponent;
-use hermes_encoding_components::traits::transform::Transformer;
 use hermes_encoding_components::traits::types::encoded::HasEncodedType;
-use hermes_protobuf_encoding_components::impls::encode_mut::proto_field::bytes::EncodeByteField;
 use hermes_protobuf_encoding_components::types::any::Any;
 use hermes_protobuf_encoding_components::types::strategy::ViaAny;
 
-#[derive(Debug, HasField)]
-pub struct WasmConsensusState {
-    pub data: Vec<u8>,
-}
-
-pub struct EncodeWasmConsensusState;
-
-delegate_components! {
-    EncodeWasmConsensusState {
-        MutEncoderComponent:
-            EncodeField<
-                symbol!("data"),
-                EncodeByteField<1>,
-            >,
-        MutDecoderComponent: DecodeFrom<
-            Self,
-            EncodeByteField<1>,
-        >,
-    }
-}
-
-impl Transformer for EncodeWasmConsensusState {
-    type From = Vec<u8>;
-
-    type To = WasmConsensusState;
-
-    fn transform(data: Vec<u8>) -> WasmConsensusState {
-        WasmConsensusState { data }
-    }
-}
+use crate::types::consensus_state::WasmConsensusState;
 
 pub struct EncodeViaWasmConsensusState;
 pub struct DecodeViaWasmConsensusState;
