+++ crates/wasm/wasm-encoding-components/src/impls/strategies/mod.rs
@@ -1,0 +1,2 @@
+pub mod client_state;
+pub mod consensus_state;
