+++ crates/wasm/wasm-encoding-components/src/lib.rs
@@ -1,1 +1,3 @@
-
+pub mod components;
+pub mod impls;
+pub mod types;
