+++ crates/wasm/wasm-encoding-components/src/types/client_state.rs
@@ -1,0 +1,9 @@
+use cgp::prelude::*;
+use ibc::core::client::types::Height;
+
+#[derive(Clone, Debug, HasField, PartialEq, Eq)]
+pub struct WasmClientState {
+    pub data: Vec<u8>,
+    pub checksum: Vec<u8>,
+    pub latest_height: Height,
+}
