+++ crates/wasm/wasm-encoding-components/src/types/consensus_state.rs
@@ -1,0 +1,6 @@
+use cgp::prelude::*;
+
+#[derive(Debug, HasField)]
+pub struct WasmConsensusState {
+    pub data: Vec<u8>,
+}
